Add v7.4 parameters file, revert input file paths to long form
As it turns out, the use of the long path (/restricted/projectnb/) or the shorter form (/rprojectnb/) does make a difference: when parsing the parameters file, GSEA 2.2.1 reads them into an array whose order is arbitrary and depends on the exact character strings in the input; it must be some artifact of hashing the strings.  Anyways, to keep things backwards-compatible, the paths are being reverted.

Also, MSigDB v7.4 just came out, so a parameters file is being added for that too.